requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
langchain==0.2.14
langchain-core>=0.2.32,<0.3.0
langchain-google-genai==1.0.8
google-generativeai>=0.5.0
python-dotenv==1.0.0
email-validator==2.1.0
aiosmtplib==3.0.1
